jh: adjust server name
The server name now includes namespace, project and commit-sha.
The pod and docker container names are of the form:

jupyter-<username>-<namespace>-<project>-<commit-sha>

"namespace" and "project" are limited to 10 characters,
"commit-sha" is limited to 7. This leaves 28 characters for
username to still yield a valid DNS name for k8s.