feat: run GH actions acceptance tests through Helm test
The CI acceptance tests are ran through Helm test. As the
tests run in the deployment cluster (rather than the GH action
container) if any test fails the results are packaged, passed
into an intermediate S3 bucket in SWITCH engine and then
downloaded in the GH action to be served as an artifact as
usual.

Fix #1762.