(feat) Build and push container images
Create CD job to build and publish the container images
for acceptance-tests, init-realm and certificates when
they change any time a commit is merged in the master
branch.

This CD job allows to deploy the Renku chart directly
from this Git repository (e.g. skipping the chart repository)
for those instances that want to deploy the tip of the
master branch of Renku (e.g. dev.renku.ch).

Part of swissdatasciencecenter/terraform-renku#139.