(feat) Publish Helm chart on all merges to master.
When a PR is merged to the master branch a chart is also published,
the version is set to `x.y.z+1-n+1` where `x.y.z` is the latest stable
release of Renku and `n` is the latest version created by
the job introduced in this PR. If an earlier pre-release
version does not exists then it is set to 0.

The workflow also pushes the tag of the just created pre-release,
this is necessary for the following workflows to assign
the correct, increasing, pre-release number.

The new workflow duplicates somewhat `publish-deploy-staging`,
however, at a later time I will take care of consolidating
the two, when Staging will also be deployed through Flux.