feat(actions) Add action to test Renku-dev
This PR is part of a work to redefine the way acceptance-tests
are piloted in production clusters: instead of having Flux running
the full test-suite, we run it from Github actions with the goal
of making the logs and the test artifacts more easily accessible and
visible.

The `pull-request-test` action is refactored to properly use the latest
version of the renku-actions.