fix(netpol): modify postgres network policy to add prometheus access (#2659)
* fix(network-policy): modify postgres network policy to permit access from prometheus

* fix(netpol): allow access to prometheus metrics from any pod in any namespace

Given that the specific namespace and labels which identify prometheus can differ from deployment to deployment, we decided it made more sense to have an open access policy on the postgresql metrics published on port 9187. Note that access to the db itself on port 5432 is still limited by this policy.