Add browser cookies to the request for getting graph status
The endpoint must now be used with an authenticated user, and since
the tests go through the UI server/Gatewa a gitlab token cannot be
used. The solution here is to grab the cookies from the browser and
add them to the request, as it is currently always called when the
testing user is authenticated.