feat: Make Redis and OIDC client secrets optional.
These are the last two secrets that were needed to be generated
out-of-bands before deploying Renkulab.

Thus the `generate-values` script is not really necessary anymore
and in its place there is a `minimal-deployment` values file that
in a future commit will also be used as a base for CI deployments.

The documentation has been updated to reflect the removal of the
script.

Further cleanup and simplification of the Helm chart will result
in a smaller and simpler minimum-deployment file, with the ultimate
goal of being able to issue `helm install renku/renku` resulting
in a functioning development deployment.