feat(helm): Make Jena passwords optional in values file.
This commit continues on the topic of making the definition
of the secrets in the values files optional, to simplify CI
deployments.

As previous similar commits this change supports:

- Existing deployments where the secret is defined will
keep working as usual

- New deployments where the secret is defined will keep
working as they did before this change

- New deployments where the secret is NOT defined will
get 2 random passwords which will be persisted across
upgrades/redeployments

The three scenarios have been manually tested.