fix: bump csi rclone to 0.3.1 (#3759)
This fixes an error where the new version of the csi rclone was failing
when a secret annotation was not used to define the secret. This was
supposed to work but the storage class that is marked to use the PVC
annotation will fail if the annotation is not present. So we had to use
2 storage classes one that requires a PVC annotation for the secret and
another that is the same as the current one - which expects that the PVC
name and secret is the same.