fix: increase timeout for authz k8s probes
The authz deployment was in a restart loop because the liveness checks
was failing with a timeout. The default is 1s and that did not work. The
period is left the same as the default but I added it just so that we
make sure we do not get into a situation where the period is lower than
the timeout.