fix: increase timeout for Authzed DB health probes (#3800)
The timeout in k8s was increasesd in a recent PR. But this increases the
timeout of the CLI we call in the Authzed DB image to actuall do the
health checks. And since this CLI had a default timeout of 1s it was
timing out even before the k8s timeout could take effect. We noticed
that Authzed on renkulab.io was in a restart loop because of this.